Some small QOL features:
The pause and play button toggle instantly, instead of waiting for the autmated background serice to update their new state.
Removed the "Coming Soon" text from the texta area card type.
Made the text area card save it's content so it can be re added when the page is reloaded.
Made it so it doesnt search for an album cover's url if there is no song playing as well as set the album cover to an empty image.
The sidebar now remembers if it was closed or not between page navigation.